correccion observaciones Manuel
